machine_learning/requirements.txt

pandas
numpy
pyserial
pyqt5
pyqtgraph
tensorflow
